+++ GGCoordinate.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GGQuadCoordinate", "GGQuadCoordinate\GGQuadCoordinate.csproj", "{07EFE6EF-4982-41FC-A06C-E2ACCD74C5A8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GGHexCoordinate", "GGHexCoordinate\GGHexCoordinate.csproj", "{89B0E7D0-69BF-4957-BBDC-607D5CC5E52A}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{07EFE6EF-4982-41FC-A06C-E2ACCD74C5A8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{07EFE6EF-4982-41FC-A06C-E2ACCD74C5A8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{07EFE6EF-4982-41FC-A06C-E2ACCD74C5A8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{07EFE6EF-4982-41FC-A06C-E2ACCD74C5A8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{89B0E7D0-69BF-4957-BBDC-607D5CC5E52A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{89B0E7D0-69BF-4957-BBDC-607D5CC5E52A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{89B0E7D0-69BF-4957-BBDC-607D5CC5E52A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{89B0E7D0-69BF-4957-BBDC-607D5CC5E52A}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
