--- GGHexCoordinate.sln
@@ -1,16 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GGHexCoordinate", "GGHexCoordinate.csproj", "{971DC953-9EA7-449B-899F-DBF05FAD963B}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{971DC953-9EA7-449B-899F-DBF05FAD963B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{971DC953-9EA7-449B-899F-DBF05FAD963B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{971DC953-9EA7-449B-899F-DBF05FAD963B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{971DC953-9EA7-449B-899F-DBF05FAD963B}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
