+++ GGHexCoordinate/FlatTop/Neighbors.iterator.cs
@@ -1,4 +1,4 @@
-﻿namespace GGHexCoordinate
+﻿namespace GGHexCoordinate.FlatTop
 {
     public static partial class Neighbors
     {
