+++ GGHexCoordinate/FlatTop/Offset.cs
@@ -1,7 +1,8 @@
-﻿using static GGHexCoordinate.Neighbors;
-using static GGHexCoordinate.Neighbors.HexagonalAxis;
+﻿using System;
+using static GGHexCoordinate.FlatTop.Neighbors;
+using static GGHexCoordinate.FlatTop.Neighbors.HexagonalAxis;
 
-namespace GGHexCoordinate
+namespace GGHexCoordinate.FlatTop
 {
     /// <summary>
     ///     "Odd Q" flat top
